services/order-api/requirements.txt
grpcio>=1.62.0
grpcio-tools>=1.62.0
psycopg[binary]>=3.1.0
psycopg-pool>=3.1.0
redis>=5.0.0
prometheus-client>=0.24.1
